Selection Sort
Added the selection sort method. Takes in an int array and spits out a
sorted version. Also added three different tests for the method
including a normal test with about 5 values, a large array test with 10+
values and a duplicate numbers test that checks to see if the method is
able to sort when there are multiple numbers in the array.